% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/center_and_scale.R
\name{center_and_scale}
\alias{center_and_scale}
\title{center_and_scale}
\usage{
center_and_scale(tbl)
}
\arguments{
\item{tbl}{- data table to centered and scaled (by column)}
}
\value{
centered and scaled data frame
}
\description{
This function filters omics data by signiificant non-parametric associations between omic features and  clinical/demographic variables.
}
