% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_data.R
\name{filter_data}
\alias{filter_data}
\title{filter_data for deploying association_wrapper}
\usage{
filter_data(
  dat,
  clinical,
  padjust = "fdr",
  sig_val = 0.05,
  filter_function = "non-parametric",
  parallel = FALSE,
  cores = NULL
)
}
\arguments{
\item{dat}{omics data, preprocesssed based on the type of dataset. Rownames are SIDs}

\item{padjust}{function for adjusting nominal p values for multiple comparisons.}

\item{filter_function}{function to be used in filtering. Currently only option in non_parametric.}

\item{parallel}{logical variable to determine whether to implement parallel processing. Implement if possible due to the memory
strain of these variables}

\item{cores}{define core numbers. Otherwise it defaults to 2 less thans  available cores}

\item{clin}{clinical features to be used in the filtering. Rownames are SIDS.}
}
\value{
subsetted filtered data frame
}
\description{
This function filters omics data by signiificant non-parametric associations between omic features and  clinical/demographic variabl
}
