% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_adjust.R
\name{linear_adjust}
\alias{linear_adjust}
\title{Regression adjustment funcion}
\usage{
linear_adjust(dat, clin, vars)
}
\arguments{
\item{dat}{omics data, preprocesssed based on the type of dataset. Rownames are SIDs}

\item{clin}{clinical data. Rownames are SIDS.}

\item{vars}{variables to regress out}
}
\value{
list of 2 dataframes, the adjusted feature data and the reduced clinical data
}
\description{
This function adjusts features for effects of variables with linear regression
}
