% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/non_parametric_assoc_wrapper.R
\name{non_parametric_assoc_wrapper}
\alias{non_parametric_assoc_wrapper}
\title{Non-Parametric Association Wrapper Function}
\usage{
non_parametric_assoc_wrapper(dat, clin, padjust = "fdr", sig_val = 0.05)
}
\arguments{
\item{dat}{omics data, preprocesssed based on the type of dataset. Rownames are SIDs}

\item{clin}{clinical data. Rownames are SIDS.}

\item{padjust}{function for adjusting nominal p values for multiple comparisons using p.adjust() function from R package Stats. 
Defaults to FDR.}

\item{sig_val}{value for significance after correction for multiple comparisons. Defaults to .05.}
}
\value{
association results for each feature
}
\description{
This function filters omics data by signiificant non-parametric associations between omic features and  clinical/demographic variables.
}
